move replaceNullWithUndefined to data client lib
