+++ libs/state-machine/src/lib/utility.ts
@@ -1,41 +1,6 @@
 import type { FormDataRecord } from '@newsletters-nx/newsletters-data-client';
 import type { ZodIssue } from 'zod';
-import type {
-	FailureDetails,
-	WizardFormData,
-	WizardStepData,
-	WizardStepLayout,
-} from './types';
-
-type PrimitiveRecordWithNull = Partial<
-	Record<string, string | number | boolean | null>
->;
-const isPrimitiveRecordAllowingNull = (
-	value: unknown,
-): value is PrimitiveRecordWithNull => {
-	if (!value || typeof value !== 'object') {
-		return false;
-	}
-	if (Array.isArray(value)) {
-		return false;
-	}
-	return Object.values(value).every(
-		(propertyValue) =>
-			typeof propertyValue === 'boolean' ||
-			typeof propertyValue === 'number' ||
-			typeof propertyValue === 'string' ||
-			(typeof propertyValue === 'object' && !propertyValue),
-	);
-};
-
-const isArrayOfPrimitiveRecordsAllowingNull = (
-	value: unknown,
-): value is PrimitiveRecordWithNull[] => {
-	if (!Array.isArray(value)) {
-		return false;
-	}
-	return value.every(isPrimitiveRecordAllowingNull);
-};
+import type { FailureDetails, WizardStepData, WizardStepLayout } from './types';
 
 export const makeStepDataWithErrorMessage = (
 	errorMessage: string,
@@ -76,38 +41,4 @@ export const validateIncomingFormData = (
 	}
 
 	return undefined;
-};
-
-/** recursively replace any `null` with `undefined` */
-export const replaceNullWithUndefined = (
-	formData: WizardFormData,
-): WizardFormData => {
-	Object.keys(formData).forEach((key) => {
-		const value = formData[key];
-
-		if (value === null) {
-			formData[key] = undefined;
-		}
-
-		if (isPrimitiveRecordAllowingNull(value)) {
-			Object.keys(value).forEach((nestedkey) => {
-				const nestedValue = value[nestedkey];
-				if ((nestedValue as unknown) === null) {
-					value[nestedkey] = undefined;
-				}
-			});
-		}
-
-		if (isArrayOfPrimitiveRecordsAllowingNull(value)) {
-			value.forEach((objectInArray) => {
-				Object.keys(objectInArray).forEach((keyOfObjectInArray) => {
-					const valueOfObjectInArray = objectInArray[keyOfObjectInArray];
-					if ((valueOfObjectInArray as unknown) === null) {
-						objectInArray[keyOfObjectInArray] = undefined;
-					}
-				});
-			});
-		}
-	});
-	return formData;
 };
